Remove withLinks and LinksProvider wrappers
